+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/bulk/imprt/model/BulkImportJobStatusEnum.java
@@ -1,4 +1,4 @@
-package ca.uhn.fhir.jpa.bulk.imp.model;
+package ca.uhn.fhir.jpa.bulk.imprt.model;
 
 /*-
  * #%L
@@ -20,25 +20,12 @@ package ca.uhn.fhir.jpa.bulk.imp.model;
  * #L%
  */
 
-import java.io.Serializable;
+public enum BulkImportJobStatusEnum {
 
-public class RawBulkImportRecord implements Serializable {
+	STAGING,
+	READY,
+	RUNNING,
+	COMPLETE,
+	ERROR
 
-	private static final long serialVersionUID = 1L;
-
-	private final String myTenantName;
-	private final String myRowContent;
-
-	public RawBulkImportRecord(String theTenantName, String theRowContent) {
-		myTenantName = theTenantName;
-		myRowContent = theRowContent;
-	}
-
-	public String getTenantName() {
-		return myTenantName;
-	}
-
-	public String getRowContent() {
-		return myRowContent;
-	}
 }
