+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/bulk/imprt/model/ParsedBulkImportRecord.java
@@ -1,4 +1,4 @@
-package ca.uhn.fhir.jpa.bulk.imp.model;
+package ca.uhn.fhir.jpa.bulk.imprt.model;
 
 /*-
  * #%L
@@ -20,12 +20,27 @@ package ca.uhn.fhir.jpa.bulk.imp.model;
  * #L%
  */
 
-public enum BulkImportJobStatusEnum {
+import org.hl7.fhir.instance.model.api.IBaseResource;
 
-	STAGING,
-	READY,
-	RUNNING,
-	COMPLETE,
-	ERROR
+import java.io.Serializable;
 
+public class ParsedBulkImportRecord implements Serializable {
+
+	private static final long serialVersionUID = 1L;
+
+	private final String myTenantName;
+	private final IBaseResource myRowContent;
+
+	public ParsedBulkImportRecord(String theTenantName, IBaseResource theRowContent) {
+		myTenantName = theTenantName;
+		myRowContent = theRowContent;
+	}
+
+	public String getTenantName() {
+		return myTenantName;
+	}
+
+	public IBaseResource getRowContent() {
+		return myRowContent;
+	}
 }
