+++ hapi-fhir-jpaserver-migrate/src/main/java/ca/uhn/fhir/jpa/migrate/tasks/HapiFhirJpaMigrationTasks.java
@@ -104,7 +104,7 @@ public class HapiFhirJpaMigrationTasks extends BaseMigrationTasks<VersionEnum> {
 		blkImportJobTable.addColumn("FILE_COUNT").nonNullable().type(ColumnTypeEnum.INT);
 		blkImportJobTable.addColumn("ROW_PROCESSING_MODE").nonNullable().type(ColumnTypeEnum.STRING, 20);
 		blkImportJobTable.addColumn("BATCH_SIZE").nonNullable().type(ColumnTypeEnum.INT);
-		version.addIdGenerator("20210406.2", "SEQ_BLKIMJOB_PID");
+		version.addIdGenerator("20210410.2", "SEQ_BLKIMJOB_PID");
 
 		// Bulk Import Job File
 		Builder.BuilderAddTableByColumns blkImportJobFileTable = version.addTableByColumns("20210410.3", "HFJ_BLK_IMPORT_JOBFILE", "PID");
@@ -113,8 +113,8 @@ public class HapiFhirJpaMigrationTasks extends BaseMigrationTasks<VersionEnum> {
 		blkImportJobFileTable.addColumn("JOB_CONTENTS").nonNullable().type(ColumnTypeEnum.BLOB);
 		blkImportJobFileTable.addColumn("FILE_SEQ").nonNullable().type(ColumnTypeEnum.INT);
 		blkImportJobFileTable.addColumn("TENANT_NAME").nullable().type(ColumnTypeEnum.STRING, 200);
-		blkImportJobFileTable.addForeignKey("20210406.4", "FK_BLKIMJOBFILE_JOB").toColumn("JOB_PID").references("HFJ_BLK_IMPORT_JOB", "PID");
-		version.addIdGenerator("20210406.5", "SEQ_BLKIMJOBFILE_PID");
+		blkImportJobFileTable.addForeignKey("20210410.4", "FK_BLKIMJOBFILE_JOB").toColumn("JOB_PID").references("HFJ_BLK_IMPORT_JOB", "PID");
+		version.addIdGenerator("20210410.5", "SEQ_BLKIMJOBFILE_PID");
 	}
 
 	private void init530() {
